add restriction page

+++ todolist_app/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render,redirect
 from django.http import HttpResponse
 from .models import Task
 from .forms import TaskForm
+from django.contrib import messages
 from django.core.paginator import Paginator
 from django.contrib.auth.decorators import login_required
 
@@ -24,7 +25,10 @@ def todolist(request):
 @login_required        
 def delete_task(request,task_id):
     task = Task.objects.get(pk=task_id)
-    task.delete()
+    if task.manage == request.user:
+        task.delete()
+    else:
+        messages.error(request,("Access Denied.!!!"))
     return redirect('todolist')
 @login_required
 def edit_task(request,task_id):
@@ -40,8 +44,11 @@ def edit_task(request,task_id):
 @login_required
 def complete_task(request,task_id):
     task = Task.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.manage == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request,("Access Denied.!!!"))
     return redirect('todolist')
 @login_required
 def pending_task(request,task_id):
